ecdlp: refactor into folder
